+++ tests/test_fw.py
@@ -6,7 +6,7 @@ import time
 import pytest
 
 from freewili.fw import FileMap, FreeWili
-from freewili.serial_util import FreeWiliProcessorType
+from freewili.fw_serial import FreeWiliProcessorType
 
 
 def test_file_mappings() -> None:
